Layouts : Improve loading of layouts with unavailable editors
This uses proxies for missing modules and editors, so we can load the rest of the layout successfully. The proxy editor appears as a tab in the UI as normal but with an "editor not available" label rather than the usual contents. The proxy can also be resaved into a new layout and then be loaded as the _real_ editor later when that becomes available.

This will allow us to load layouts from Gaffer 1.4 in 1.3, even if they include an ImageInspector or LocalJobs panel. Saving the layout in 1.3 and reloading it in 1.4 will even preserve the new editors.