CI : Add `linux-gcc11` build variants
Update to new build containers with `GAFFER_BUILD_ENVIRONMENT` environment variable. We can use this to download the appropriate dependencies package for each build environment.

We could potentially use `GAFFER_BUILD_ENVIRONMENT` to also define the build name in setBuildVars.py. This would need to be some combination of the platform, build type (nothing for a release build, "-debug" for a debug build), and the build environment.