ImagePlug : Allow serialisation of child connections/values
This means we also need to disable serialisation for `ImageProcessor.out` to avoid unwanted serialisation of internal connections. This _could_ have been achieved by removing the Serialisable flag, but that caused problems for things like `Dot::setup()`, which _do_ force the added plug to be serialisable, but _not_ its children. We could have modified all those locations to set flags recursively, but since we want to remove flags entirely in future it makes more sense to use a custom serialiser instead. The Dot setup problem demonstrates the problem with flags in the first place - they get copied from place to place, but they may only have made sense on the node they originated from.

In conjunction with the previous commit, this fixes #3986.