OSLShader : Always load output parameters in `loadShader()`
Prior to this, and dating all the way back to the inception of GafferOSL,
we were only loading output parameters when the shader type was `shader`,
the generic shader type. This was incompatible with 3Delight's shader library
which declares `surface` shaders to have a `closure outColor` output parameter
to allow them to be connected as inputs to things like `dlLayeredMaterial`.

Considering the OSL ecosystem more broadly, the OSL project itself has said
that [shader types are deprecated](https://github.com/AcademySoftwareFoundation/OpenShadingLanguage/pull/899)
and all shaders are created equal - "Generic `shader` type is the only one
we will talk about as we move forward". Always loading all output parameters
for all shader types is also consistent with this direction of travel.

This is a breaking change, so it does come with some compatibility concerns,
but the new test case demonstrates that old scripts will still load and
that old connections to ShaderAssignment nodes continue to work as before.

This work is in response to a comment on https://github.com/GafferHQ/gaffer/pull/5617 :

> In Gaffer currently OSL shaders with closure input parameters
> (for ex. 3Delight's Layered Material shader) don't accept connections
> from closure output parameters of OSL surface shaders, only from OSL
> generic shaders. Due to this it's not possible to connect any of the
> default surface shaders that come with 3Delight (dlPrincipled, dlGlass,
> dlMetal, dlStandard, etc.) to the closure inputs of dlLayeredMaterial or
> dlRandomMaterial.