LocalDispatcher : Don't execute in background from `execute` app
As noted in the comments, in this circumstance the job would always have been killed before it could complete. That wouldn't ever be anybody's intent, but it was quite an easy situation to get into accidentally, because the `userDefault` for `executeInBackground` is `True`. We force foreground execution rather than treat this as an error because when testing nested dispatcher setups you'd want background execution on when testing the inner dispatcher in isolation, and would want it off automatically when launching the outer dispatcher.

Note that it's still possible to launch multiple background jobs from a Wedge by ensuring that the outer dispatcher has `executeInBackground = False`. And that testing for the `execute` app works equally well when the outer dispatcher is a TractorDispatcher.