GafferCycles startup : Add basic back-compatibility for `principled_bsdf`
This has breaking changes going in to Blender 4. Here we're just dealing with the renamed parameters, which is sufficient to get old scripts loading. But some parameters have also changed semantics, in such a way that they can only be emulated by inserting new shaders into the graph - they will need to be dealt with later.