GafferSceneTest : Avoid deprecation warning from `random` module
This fixes test failures due to the following warning : "Seeding based on hashing is deprecated since Python 3.9". Removing the `datetime` argument leaves us in a similiar situation to before, because a default-contructed `Random` instance seeds from "current time or from an operating system specific randomness source if available" anyway.

I do think the original intent is questionable : a test that uses different random numbers each time it runs is going to be infuriating if it fails, because we'll have no way of reproducing the problematic input. But that can wait - right now we're just making the minimum changes necessary to upgrade to  Python 3.10.