ScriptNode : Remove AST code
The API we were using to do this is no longer available in Python 3.10, and wasn't available on Windows or MacOS anyway. This does lead to around a 7% performance regression when loading a large template from `gaffer-examples`, but there's little we can do about it without changing approach completely.

Alternatives tried at this point are :

- Continuing to use the AST, accessing it via the Python `ast` module rather than the C API. This led to a 80% regression.
- Doing our own parsing for the very long `__children["child"]["child"]["child"]` prefixes present on most lines of the serialisation. This was no better, because we still need to use `__getitem__` at each step of the way, and it seems that is where much of the overhead is. We can't stop calling `__getitem__`
because our monkey-patching compatibility configs rely on it. If we could, then we could move a great deal of the execution to pure C++, which I think might be a big win.