RenderPassEditor : Add render pass deletion
We allow deletion of render passes created within the same Edit Scope by removing the names of deleted render passes from the list of names created within that Edit Scope.

If a user attempts to delete a render pass that was created in an upstream Edit Scope, we prompt them to either disable those render passes in this Edit Scope or manually switch to the upstream Edit Scope and delete them there.

Attempts to delete a render pass that was created in a downstream Edit Scope are met with a dialogue suggesting the deletion should occur downstream but with no option to disable the render passes in order to limit the creation of upstream edits to downstream render passes.