IECoreCycles : Remove `devices()` method
This was just the same as Cycles' own `available_devices()` method, but with the additional of a strange "MULTI" device that included every single other device, including both the CUDA and Optix variants of the same NVIDIA card. I don't know what that device was intended for, but it doesn't seem to make sense.
While "MULTI" does seem to be used as a value for `CyclesRenderer::m_deviceName`, in that case it seems to mean something else - to use the list of devices gathered into `m_multiDevices`.

The one potential downside here is that we now call `available_devices()` more, but internally it keeps track of what has been initialised already to avoid repeat work anyway. And most of those calls are going to disappear in upcoming refactoring anyway.