gaffer.cmd : Disable OIIO `add_dll_directory`
Python changed the paths it searches for module binaries in version 3.8.
It now does not search the `PATH` environment variable. OpenImageIO
works around this by adding all directories found in `PATH` to the
Python search list. But Python throws an exception when trying to add
directories that start with a `..` relative directory. Besides that,
not adding everything from `PATH` is meant to be more secure, so this
seems like a reasonable thing to disable.