GafferCycles : Don't nest `__ParametersPlugProxy`
This fixes a bug in `principledBSDFCompatibility` that was preventing
use of the 3d Viewer. The error was
`IECore.Exception: Line 276 of startup\gui\cyclesViewerSettings.gfr :
 AttributeError: '__ParametersPlugProxy' object has no attribute
 'parent'`

Every time we were calling `GafferCycles.CyclesShader.__getItem__` on
a plug named `parameters`, we were putting that plug into a
`__ParametersPlugProxy` and returning that object. This meant on the
second `__getItem__` we'd be putting a `__ParametersPlugProxy` inside a
`__ParametersPlugProxy` and it would error because
`__ParameterPlugProxy` doesn't have a `parent()` method (called on L60)