Context : Fix heap read after free
This was introduced in 520c7a35d8818dba0a9f37bd36fbf8ce3dd2678d, and detected by running `ContextTest` and/or `ScriptNodeTest` using an ASAN-instrumented build. The issue was that `internalSet()` compares the new value against the (non-owning) old value stored in `m_map`, but the memory that points to has been deleted already by the assignment to `m_allocMap` (because assignment to `intrusive_ptr` decrements the reference count for the previous pointee). The solution is to temporarily hold on to the old value until after calling `internalSet()`, and to do all the `intrusive_ptr` juggling using `std::move()` so we steal references rather than do unnecessary increment/decrement pairs.

The one other place we call `internalSet()` while using `m_allocMap` is in the Context copy constructor, where we are OK, because `m_allocMap` starts out empty.