Switch : Add `deleteContextVariables` plug
This provides a workaround for the fact that DeleteContextVariables can't currently be used on a variable-length ArrayPlug input, because there is nothing to keep it in sync with the length of the ArrayPlug output. It's a valuable workaround because it's not uncommon to use a context variable to provide the switch index, and we can cleaner upstream contexts if we can delete such variables.

I think it probably would be possible to get DeleteContextVariables to sync the lengths of its input and output. But the cleanest way of doing so is probably to do it in `Plug::parentChanged()` where we're already doing the same for plugs associated by connections. We would need some way of telling Plug that there are other "paired" plugs that should be synchronised despite not being connected. I'm not sure what that should look like, and it might well be an ABI break. So for now I'm taking the cheap and cheerful route.