Cycles Renderer : Remove misleading member data
It was unclear what the "source of truth" was here - the nodes in the `ccl::Scene` or the member data. It looks like the member data was always just being used to store a copy of the scene data so that we could restore it after doing a scene reset. We don't need member data for that - we can just use temporary variables in `reset()` itself. The less member state we have to track, the better.