CyclesOptions, CyclesRenderer : Remove `useFrameAsSeed`
This didn't map to an actual Cycles parameter, and made for a more complex mechanism than the one we use for Arnold. Behaviour now matches Arnold, including using the correct default value for the frame.

We don't expose the sampling seed for 3Delight yet, but this "use explicit seed or fall back to frame" behaviour will map very well to it in future. This might even make a good candidate for promoting to a standard option.