CyclesRenderer : Remove `m_*ParamsDefault`
And replace with functions to get the defaults whenever we want them. This does add a tiny bit of overhead to `option()` but I doubt it's measurable, and `option()` isn't performance-critical anyway. In return we get to remove state from Renderer, and set ourselves up a bit better for the upcoming session-reset-refactor.