Cycles Renderer : Remove strange film/integrator update mismatch
The call to `film->tag_update()` was commented out in d78e359a5258d0e568f972031cebdff51081eb08, a "Cycles 2.93 API update" commit containing a huge number of other changes. It looks like the `Film::tag_update()` method was removed in https://github.com/blender/cycles/commit/588eb94148e3ea62457b666496cbfc3919c44860. And the reason for that seems to be that `Film::tag_update()` was no longer necessary because it could track it's own dirtiness manually using the new modified flags managed by the Socket API.

But what's less clear is why we should call `Integrator::tag_update()` if the film has changed. I can't see any reason for that. In the absence of clear evidence for existence, we can't keep code like this hanging around forever, so I'm tearing off the band-aid now. Added a unit tests that seems to show that this is fine.