Renderer : Allow null attributes in `camera()` call
This is a concession to help the Cycles backend accomodate the needs of RenderController and SceneGadget, which both find it easier to create a default camera before outputting all the options. Although it is pleasing to see that it also simplifies the vast majority of calls to `camera()` in our unit tests.