Renderer : Add todo
If we made this modification to the Renderer API, the Cycles backend would have more options in dealing with the options-before-session but session-before-attributes chicken-and-egg problem. If `Renderer::attributes()` was called before the session was acquired, it could have just returned a special `PendingAttributes` object that stored the source CompoundObject. Then later in `Renderer::editAttributes()` after the session has been made, it could have transformed PendingAttributes into the final CyclesAttributes form.

This would potentially also have benefits in the 3Delight backend, meaning that DelightObject didn't need to store a DelightContext.