Widget : Offset `ButtonEvent.line` by 0.5 pixel
The line representing the pixel location of a button press should go
through the center of the pixel, not the corner. This was causing
problems with getting precise object hit locations in `SceneGadget`.

Interactive testing with a variety of widget types suggests this hasn't
made anything in the UI respond in unexpected ways to mouse clicks.