Pointer : Remove use of `IECoreImage::ImagePrimitive`
We want to phase ImagePrimitive out over time, and this is a small step in that direction. This does leave us only being able to create ImagePrimitives from filenames, but that is the only constructor we've ever used anyway.

If we did want to construct from arbitrary in-memory data, I'm not sure what we'd choose. We wouldn't want to use `QPixmap` because the goal of GafferUI is to entirely hide Qt from the public API. So maybe `OIIO::ImageBuf`? That's for another day though since we don't need it at the moment.