Render : Support `render:defaultRenderer` option
The fact that we now compute the globals in `hash()` might be a bit of a concern for folks who have made their scene globals slow to compute for some reason. But it's the logical thing to do, and it also has the potential to improve the hash result in future (see todo).