InteractiveRender : Support `render:defaultRenderer` option
It would be nice to be able to support live edits to this, switching the currently running renderer on the fly. Technically I think this could be done relatively simply by storing the current renderer type when making a renderer, then checking to see if we should recreate when the globals are dirtied. We'd need an ABI break to store that though. And I think there is a better approach that we should take in the longer term : let the RenderController create and recreate the renderers itself. That would allow us to keep the globals evaluations in a background thread for the viewer, and would also let renderers reject certain option edits (e.g. `cycles:device`), forcing the controller to create a new renderer. That would be a bigger API/ABI break, so I'm not tackling it now, because we're targeting `1.3_maintenance` with this work.