RenderPassEditor : Add `render:defaultRenderer` column
This is in desperate need of custom widgets for editing options, so we can provide a presets menu. But I think that will be better tackled more holistically, considering the requirements of other columns too. The current goal is to get the basic functionality in quickly to avoid Cinesite having to roll their own.