PythonExpressionEngine : Support compatibility configs for renamed plugs
The key change here is that `__plug()` now uses a series of `__getitem__` calls rather than a single call to `descendant()`, giving the compatibility configs the ability to monkey-patch a name change in at any step.

The other changes aren't strictly necessary, but I switched from storing plug paths as "dot.separated.strings" to `( "tuples", "of", "names" )` since that's the form we now need them in most commonly, and avoids a bunch of string manipulations for the more common operations. This leaves us doing string manipulations only for the less common error-handling code paths.