USDShader : Load lights from UsdSchemaRegistry
At this point, this should be giving exactly the same results as loading them from the SdrRegistry (which is populated from the schemas in the first place). The reason for changing is that the SdrRegistry intentionally ignores auto-apply schemas used to add renderer-specific properties to lights, and we don't want to ignore them. By loading from the schemas directly ourselves, we'll be able to use auto-apply schemas in an upcoming commit.

It may be that we'll be able to load everything (including regular shaders) from schemas in future, according to a [comment from spiffmon](https://github.com/PixarAnimationStudios/OpenUSD/pull/2497#issuecomment-1608252125) :

> I'd like to really explore a future where Sdr is primarily a rendering-focused subsystem...but enable USD clients/tools to remain unaware of it. In other words, push forward on what we started in our integration of RenderMan into USD, where `usdgenschemafromsdr` can be leveraged to produce schemas not just for "shader like" RenderMan objects/API's, but for shaders in general.