ImageGadget : Support 2-channel images
Although we never used them ourselves, the old TextureLoader-based code path prior to 8ea7d9f63e68759d6988afb280f464d94d742a41 did support them, and it turns out they are in use in the wild. I've used `GL_LUMINANCE_ALPHA` because that is what was used by the old code path, but technically it is deprecated - in future we'd need to use `GL_RG` and `GL_TEXTURE_SWIZZLE_RGBA` instead.

Also fixed an empty filename in the "Unsupported number of channels" error message - it seems the `ImageBuf` filename is lost during the `colorconvert()` call.