GafferUI : Fix crashes during shutdown
All these signals get connected to by Python functions, so will contain slots with a `PyObject *` in them. But Python shuts down _before_ static destructors are run, after which we can't delete the PyObject. The solution is one we've used elsewhere : just "leak" the signal and its connected slots. It's intended to live for the entire lifetime of the app anyway, so it doesn't matter if it's not cleaned up during shutdown.