Install : don't check inkscape capabilities if disableGraphics is True
There is a way to disable the generation of some graphics, like for example
when Inkscape is not available. This is done by providing the INKSCAPE
argument to the build command and setting it to “disableGraphics”.

However this was still running Inkscape, and in our case it was failing.
Changes in this commit should avoid that.