Shader : Always specify output name in ShaderNetwork connections
We were previously omitting it for subclasses which didn't use the `out` plug as a container for named shader outputs, and instead used it to represent a single output. Everything is simpler if outputs always have a name, and this aligns much better with USD too where a name is required. Cortex has a nasty workaround with a `DEFAULT_OUTPUT` to round-trip empty names though USD, and this will no longer be needed.

What we really want to do as well is enforce that the `out` plug is always just a container, and all outputs must be parented under it. This will also allow us to support multiple Arnold outputs, as requested in #5542. But maintaining backwards compatibility for that change is a bigger problem. This commit means that the `IECoreScene::ShaderNetworks` we generate and export have the format we want, even though we haven't yet adopted the plug format we want yet.

Note : the choice of `out` for the default output matches Autodesk's convention for the default (unnamed) output from an Arnold shader, so this means we now export USD using the same conventions as Maya.