ViewportGadget : Avoid rendering selection ids with `GL_BLEND` enabled
Enabling GL_BLEND for selection renders on Intel graphics hardware leads to corrupted ids in the selection buffer, so we now ensure it is disabled for selection renders in `renderInternal`.

We also consolidate the enabling of GL_BLEND so it only occurs in `renderInternal`, rather than in various parts of StandardStyle and in `ViewportGadget::render`.

This fixes #901 and #2788 on the Intel graphics hardware we have available to test with.