Delight Renderer : Set `.global.frame` parameter from "frame" option
This is used by 3Delight to vary the sampling pattern per-frame. It's not clear if this has always been broken, or if at some point the `frame` argument to `NSIRenderControl` did the same thing. The latter is no longer documented in `nsi.pdf`, but there is no mention of why in the 3Delight change log, which appears not to have been updated for a year.