GafferCyclesModule : Call `ccl::path_init()`
Because we're linking to static Cycles libraries, any static variables
will be local to that instance of the linked library. By calling
`IECoreCycles::init()`, we were registering the Cycles root directory
with the library linked to `IECoreCycles` / `GafferCycles`. That meant
`GafferCycles` _module_ was not registering the Cycles root, and
GPU devices would fail to initialize because it could not find their
binary compilers. Registering the Cycles root in both modules allows
all Cycles library instances to find the files it needs.