GraphEditor config : Use GraphGadget as drag target, not GraphEditor
Although they currently share the exact same screen real estate, that is not guaranteed in future. For instance, a toolbar might be added, and we wouldn't want to accept drags onto that.

The implementation is also slightly simpler this way, since we don't need to map from raster to gadget space to figure out where to put the nodes.