OutputBuffer : Fix unwanted overlays when selection is empty
Because we were using `[ 0 ]` as an empty selection buffer, we would draw an overlay for any non-zero alpha area with a value of zero for the ID. Two such instances were :

- An Arnold `skydome_light` which was visible to camera.
- A 3Delight render where the ID output hadn't yet updated, but the RGBAZ had.

I'm not certain of the reason for using `[ 0 ]` for empty selection buffers - just having an empty array would certainly be more elegant. But I'm reluctant to mess about changing that right now because of the possibility that it was needed for certain drivers somewhere.