Delight RendererTest : Improve NSI parsing
We were checking for an incomplete list of "{flags} {type}" combinations in order to get the "{type}" part we were interested in - omitting among other things "vl" for "{flags}" and "vector" and "color" for "{type}". Instead we can just split the string and get any type reliably in one go.

At the same time, I've used the exact type string for the "v3f-ish" types. Although not strictly necessary at the moment, I think this is clearer than "v" because in 3Delight's NSI syntax "v" means `NSIParamPerVertex`. This also moves us closer to being able to accurately test exact types in future.