Capsule, Unencapsulate : Fix cancellation handling
We weren't omitting the canceller when copying the context in the Capsule constructor. This meant the Capsule could contain a dangling pointer to a long-since destructed Canceller, which among other things could cause `IECore::Cancelled` to be thrown when unencapsulating or rendering the capsule. Even if the Canceller wasn't dangling, it still wasn't suitable to be stored, because the Capsule is stored in Gaffer's cache and could be used subsequently by a client for which the original canceller doesn't apply.

A similar bug existed for the Unencapsulate node, which needs to ensure the _current_ Canceller is used when evaluating the Capsule contents. We do this in CapsuleScope by combining the current canceller with the Context stored in the Capsule, via an EditableScope.

There was one limited scenario in which the first bug could actually have worked in our favour, and which isn't matched by the fix. When rendering a freshly-minted capsule, the canceller might have still been valid, allowing cancellation of a capsule expansion. This wouldn't have applied to any of our current rendering mechanisms though; neither batch nor interative renders use cancellation, and the Viewer doesn't expand capsules. This does point out a limitation of Capsules though (and Procedurals more generally) - we need to add a mechanism to pass cancellers to the `render()` method.