Arnold USD schemas : Put light properties in main sections
And use a label suffix to make it clear that they are Arnold-specific. It seems like this might be preferable for users of Arnold, because the parameters are closer to hand. But it probably won't scale very well as we add schemas for other renderers. I think it might be preferable though, at least in the short term, as our first big use case is trying to replace Arnold lights in an Arnold pipeline.

It feels like this could also be improved aesthetically if we used icons instead of a textual suffix.