GraphEditor config : Improve pointer for targeting dragged locations
The old pointer felt too much of a departure from the "objects being dragged" pointer, and was too big of an emphasis for what is probably a pretty niche feature. It felt like you'd stopped dragging objects, making things less clear when you dragged onto a PathFilter, which is the far more common operations. The pointer now matches the "objects being dragged one" with the addition of a discrete targeting crosshair in the lower left.