TweakPlug : Fix bugs in list modes
Passing the same data into `applyListTweak()` as both the `const tweakData` and the non-const `destData` was not a good idea. It encouraged the implementation to think they point to different objects, when in fact they do not. This was further obfuscated behind the `tweakDataCast` and `newElements` variables, both of which continued this aliasing. The result : when we write to the destination data, we're clobbering the tweak data.

It's somewhat miraculous that this worked at all, but it did in the common case, because :

- The `tweakData` came from `PlugAlgo::getValueAsData()`, which takes a copy of the data held on the plug. So we had two Data objects pointing to the same internal data (`tweakData` is one, the value held by the plug is another).
- The `const newElements` reference was taken first, referencing this same internal data.
- Then `data->writable()` was called, doing a copy-on-write of the internal data so that `data` owns a new internal object and `newElements` continues to reference the internal object owned by the data in the plug.

This fell apart when the plug _didn't_ continue to reference the original data, because then there is no copy-on-write and `newElements` and `data->writable()` are the same thing. This could happen two ways - either the plug computes an uncached value, or the value is evicted from the cache concurrently with it being used by TweakPlug. Both would be pretty horrendous to track down in a production setting.

Longer term we should replace the two function arguments with a single one to make it clearer what is happening. We can't do that until an ABI break though, because even though it isprivate, the function is part our ABI due to being called by public inline functions.