Cycles ShaderNetworkAlgo : Workaround bad connections made by ShaderView
This fixes the rendering of shader balls in the Viewer, which was broken by
https://github.com/GafferHQ/gaffer/pull/5759. This really is just a workaround though, and the ideal fix is now documented in ShaderView.cpp, along with the reasons we can't make it immediately.

I moved the check for "" and "out" after the search for the valid outputs because of the small chance of there actually being a real output called "out" (and it not being the first one). I think it's unlikely that we need the check for "" as well any more, but I've kept it out of an abundance of caution.

Also changed InternedString equality comparisons to use preconstructed strings - comparing to `char *` actually constructs InternedStrings to compare against on the fly, which is unnecessarily slow.