AttributeTweaks : Add fallback for `linkedLights` attribute value
This is necessary to make `{source}` substitutions useable in all situations. As noted in the comment, we should really do this for _all_ attributes using a central registry of default values. We're just doing this one on its own because it is such high value - `linkedLights` is the main thing we had in mind when adding the `{source}` token in the first place.

Note that once we have fallbacks for all standard and renderer-specific attributes, the `ignoreMissing` plug will only be relevant for custom attributes not present in the registry. I view this as a good thing. If I make a `Multiply` tweak to double `ai:AA_samples`, it's not because I want to double it only if it exists. Rather, it's because I want to double the value that is received by the renderer.