ContactSheet/ContactSheetCore : Add `filter` plug
This provide control over the sharpness/ringing/aliasing tradeoffs associated with resizing the input images into tiles.

Deliberately hardcoding the filter used for label/border, because they are subject to less resizing anyway, and you wouldn't want to get ringing on their high-contrast edges just because you wanted to sharpen up the images in the tiles.