HiddenFilePathFilter: Fix hidden sequences
Fixes #5810

When checking a path that has the `#` character for frame substitutions,
`HiddenFilePathFilter` would always treat it as hidden. The file with
`#` does not exist, so the Windows API would return
`INVALID_FILE_ATTRIBUTES` whose value is `-1`. `FILE_ATTRIBUTE_HIDDEN`
has a value of `2`. `-1 & 2 = 2`, so sequences would always be
considered hidden.