Delight Renderer : Register cloud rendering variant
In an ideal world perhaps, we'd have implemented this as an option specified by DelightOptions, but that would put us in the chicken-and-egg situation where we can't make an NSI context without the cloud option, and we can't handle the other options until we have an NSI context. Although there _are_ strategies for dealing with that (see CyclesRenderer) they're complex and it would be preferable to avoid them. So instead this is implemented by registering a second "3Delight Cloud" renderer type, with the justification being that cloud rendering is a different enough operation that it's reasonable to present it to a user this way. In practicals terms there is no difference between this and a potential option, because the renderer can be specified using the standard `render:renderer` option.