LightPositionTool : Default pointer on handles
When we moved to showing the `notEditable` pointer for key combinations
that are not clickable / draggable, the `notEditable` pointer would
persist on the rotate and distance handles even when they are editable.

This sets the pointer to the default when the cursor is over those
handles, leaving it to the handle color to indicate editability as is
done elsewhere.

Also switched the check for pivot editability to use `DistanceHandle::getRequiresPivot()` rather than explicit mode checks.